refactor: add 'Primitive' postfix
