+++ packages/bezier-react/src/primitives/CheckboxPrimitive/CheckboxPrimitive.test.tsx
